feat(logging): add winevtlog as another supported plugin
winevtlog is the new plugin to read windows logs from using the winevtlog
API, the old one is winlog but it works for many cases so we want to keep
both.

We though of doing some magic and auto-configure this new plugin for winlog
(so replace winlog > winevtlog when creating the fluent-bit config) but the
problem is that the outputs of those plugins are pretty different, one uses
Computer, the other ComputerName and so on.

So we've decided to set this winevtlog as default on the examples but do
not change existent the behavior of existing installations.

Not the best in the world but makes sense, also customers will be able to use
one or other without the need of creating custom fluent-bit configurations.